Module TpIndividual::condiciones

Structs

Condicion

Estruct para poder guardar las condiciones separandola en datos

Enums

OpLogico

Enum para representar los distintos tipos de operadores logicos

Operador

Enum para representar los distintos tipos de operadores

Functions

comparar_con_csv

Funcion para comparar las lineas del csv y ver si cumplen las condiciones ingresadas #Recibe por parametro un vector con las condiciones parseadas, la ruta del csv y el header -Itera sobre el csv y lo lee linea por linea -Compara si la linea del csv cumple con las condiciones que llegan por parametro -Si cumple las condiciones lo pushea a una matriz -Repite el proceso hasta terminar de recorrer el csv -Finalmente retorna una matriz con las lineas que cumplan con las condiciones

comparar_op

Funcion para comparar las operacion que tiene cada condicion #Recibe por parametro la condicion, la fila y la posicion -Itera sobre el operador de dicha condicion -Realiza la operacion entre el valor de la fila del csv y nuestra condicion -Retorna si es verdadero

comparar_op_logico

Funcion para comparar utilizando el operador logico que recibe de la consulta #Recibe por parametro las condiciones parseadas, la fila y el header -Revia que la columna de la condicion exista y devuelve su posicion -Compara la OP para observar si es verdadera o falsa -Vuelve a comparar pero usando nuestro operador logico para tener un resultado final -Devuelve el resulta final luego de realizar las comparaciones

obtener_op

Funcion para obtener el tipo de operador

obtener_op_logico

procesar_condiciones

Funcion para procesar los distintos tipos de condiciones que pueden llegar dependiendo de la consulta -Ejemplo #Recibe un vector de condiciones como por ejemplo: [“producto”, “=”, “Teclado”, “AND”, “cantidad”, “>=”, 1] -Itera sobre los elementos de este vector y va rellenando un strut “condicion” -Luego pushea la nueva condicion en un vector de condiciones parseadas -Repite e itera hasta llegar al final de la lista -Finalmente retorna el nuevo vector con las condiciones “parseadas”