Module TpIndividual::manejo_de_string

Functions

crear_matriz

Funcion para crear una matriz a la hora de utilizar el INSERT con multiples valores #Recibe por parametro el String con los valores a insertar -Realizar un trim de limitado por parentesis y por las comas -Luego mapea los valores y los pone en un vector -Finalmente los devuelve en formato de matriz #Ejemplo: -Recibo por parametro (1,2,Monitor,22), (1,3,Monitor,0) (mis valores a insertar) -Elimina los parentesis y realiza realiza un split con las comas y va juntando los valores como vector -Luego añade los vectores a un vector de vectores para formar una matriz -Finalmente retorna la matriz creada

obtener_primera_palabra

Funcion para obtener la primera palabra de nuestra consulta #Recibe una cadena por parametro con la consulta completa -Divide la cadena por espacios -Devuelve la primera palabra encontrada -En otro caso devuelve un string vacio (no hice que devuelva un error por que se maneja en el main)

separar_datos

Funcion para separar los datos de la consulta INSERT #Recibe por parametro la consulta -Define un vector con dos partes, usando VALUES para separar -Luego separa esas dos partes y opera para dejar valores y direccione_y_columnas como Strings separados -Finalmente retorna los dos Strings

separar_datos_delete

Funcion para separar los datos de la consulta DELETE #Recibe por parametro la consulta sql -Divide la consulta en dos partes una con el nombre del csv y otra con la clave -Con la primera cadena que contiene el nombre reemplaza el DELETE y el FROM para dejar solo el nombre del csv -Con la segunda cadena que contiene los valores itera sobre dicha cadena y deja solamente la calve y el valor a actualizar -Finalmente retorn un string con el nombre y un vector clave-valor -En otro caso devuelve un error

separar_datos_select

Funcion para separar los datos de la consulta SELECT #Recibe por parametro la consulta sql -Divide la consulta en dos partes una con el nombre + columnas y otra con la clave -Con la primera cadena que contiene el nombre reemplaza el SELECT y el FROM para dejar solo el nombre del csv Y en otra variable las columnas del SELECT -Con la segunda cadena que contiene las condiciones quita los ; y recoge todo en un vector. -Finalmente retorn un string con el nombre otro con las columnas y por ultimo un vector con las condiciones -En otro caso devuelve un error

separar_datos_update

Funcion para separar los datos de la consulta UPDATE #Recibe por parametro la consulta sql -Divide la consulta en dos partes una con el nombre del csv y otra con los valores -Con la primera cadena que contiene el nombre reemplaza el UPDATE para dejar solo el nombre del csv -Con la segunda cadena que contiene los valores itera sobre dicha cadena y separa la clave de los campos a actualizar -Finalmente retorn un string con el nombre, un vector con los valores y otro con la clave para actualizar -En otro caso devuelve un error

separar_order

Funcion para separar el ORDER de las condiciones de un SELECT #Recibe por parametro un vector con las condiciones de la consulta -Si contiene order itera sobre la cadena y quita el order para almacernar el resto de los strings en un vector -Separa el vector en dos strings y luego los mapea en dos vectores que contengan las condiciones y por otro lado el ORDER -Finalmente devuelve dos vectores uno con el criterio de ordenamiento y otro con las condiciones -En caso de no contener ORDER devuelve un vector con las condiciones y uno de ordenamiento vacio.